Do not abstract StorageSwitch over content
We store &Column and &ComponentSparseSet in `StorageSwitch` rather than
abstract it over content, so that usages are clearer. This also
guarantees our union size is exactly the size of a pointer. Instead of
transforming the &Column into various references to columns when
constructing `StorageSwitch`, we just store the &Column. This required
updating access to account for this change.

Seemingly, this should not be a performance drop, since the level of
indirection is the same, and we unsafe_unwrap the same way as before.